Add first lesson
Add patches and setup script for creating the conflict.

The setup script does not (yet) check if the lesson was already
finished in the main branch (and reset it to "neutral")

We also need a `check.sh` script to see if the repo is in the desired
state (both patches applied to main as merge commits).